fix: permit send decimal in price to add busdet item

+++ API/UseCases/BudgetItem/CreateBudgetItem/CreateBudgetItemUseCaseInput.cs
@@ -2,7 +2,7 @@ using Microsoft.AspNetCore.Mvc;
 
 namespace AutologApi.API.UseCases
 {
-    public record BudgetItemUseCaseInput(string Id, string Description, int Qtd, int Price);
+    public record BudgetItemUseCaseInput(string Id, string Description, int Qtd, decimal Price);
 
     public record CreateBudgetItemUseCaseInput(
         [FromBody] BudgetItemUseCaseInput BodyData,
